requirements.txt
accelerate==0.27.2
action-msgs==1.2.1
action-tutorials-interfaces==0.20.3
action-tutorials-py==0.20.3
actionlib-msgs==4.2.3
addict==2.4.0
ament-clang-format==0.12.10
ament-cmake-test==1.3.8
ament-copyright==0.12.10
ament-cppcheck==0.12.10
ament-cpplint==0.12.10
ament-flake8==0.12.10
ament-index-python==1.4.0
ament-lint==0.12.10
ament-lint-cmake==0.12.10
ament-package==0.14.0
ament-pep257==0.12.10
ament-uncrustify==0.12.10
ament-xmllint==0.12.10
angles==1.15.0
annotated-types==0.6.0
ansi2html==1.9.1
anthropic==0.19.1
anyio==3.7.1
argcomplete==3.1.6
asttokens==2.4.1
attrs==23.2.0
blickfeld-scanner==2.20.6
blinker==1.7.0
bond==3.0.2
builtin-interfaces==1.2.1
cachetools==5.3.2
can-msgs==2.0.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
clang-format==17.0.5
click==8.1.7
cloudpickle==3.0.0
cmake==3.27.7
comm==0.1.4
composition-interfaces==1.2.1
ConfigArgParse==1.7
cv-bridge==3.2.1
Cython==3.0.7
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
demo-nodes-py==0.20.3
diagnostic-aggregator==3.1.2
diagnostic-msgs==4.2.3
diagnostic-updater==3.1.2
distlib==0.3.8
distutils-extra-python==2.39.2
docker-pycreds==0.4.0
domain-coordinator==0.10.0
easyocr==1.7.1
einops==0.7.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.1
examples-rclpy-minimal-action-client==0.15.1
examples-rclpy-minimal-action-server==0.15.1
examples-rclpy-minimal-client==0.15.1
examples-rclpy-minimal-publisher==0.15.1
examples-rclpy-minimal-service==0.15.1
examples-rclpy-minimal-subscriber==0.15.1
exceptiongroup==1.1.3
executing==2.0.1
Farama-Notifications==0.0.4
fastapi==0.104.1
fastjsonschema==2.19.1
filelock==3.12.4
Flask==3.0.0
fsspec==2023.10.0
gdown==5.1.0
geodesy==1.0.5
geographic-msgs==1.0.5
geometry-msgs==4.2.3
gitdb==4.0.10
GitPython==3.1.37
grid-map-msgs==2.0.0
gymnasium==0.29.1
h11==0.14.0
hiq-python==1.1.12
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.2
identify==2.5.35
idna==3.7
ifaddr==0.2.0
image-geometry==3.2.1
imageio==2.34.0
interactive-markers==2.3.2
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kazam==1.5.7
laser-geometry==2.4.0
launch==1.0.5
launch-ros==0.19.7
launch-testing==1.0.5
launch-testing-ros==0.19.7
launch-xml==1.0.5
launch-yaml==1.0.5
lazy_loader==0.3
lifecycle-msgs==1.2.1
logging-demo==0.20.3
map-msgs==2.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
message-filters==4.3.3
mpmath==1.3.0
nav-msgs==4.2.3
nav2-common==1.1.13
nav2-msgs==1.1.13
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
ninja==1.11.1.1
nmea-msgs==2.0.0
nodeenv==1.8.0
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
open3d==0.18.0
opencv-python-headless==4.9.0.80
osrf-pycommon==2.0.2
packaging==24.0
pacmod3-msgs==1.0.0
pandas==2.1.2
parso==0.8.3
pathtools==0.1.2
pcl-msgs==1.0.0
pendulum-msgs==0.20.3
pillow==10.2.0
pipx==1.2.1
platformdirs==3.11.0
plotly==5.18.0
pre-commit==3.6.2
prompt-toolkit==3.0.39
psycopg2==2.9.9
pure-eval==0.2.2
py-itree==0.0.19
pyclipper==1.3.0.post5
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.16.1
pyquaternion==0.9.9
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-qt-binding==1.1.2
python-xlib==0.33
PyYAML==6.0.1
pyzed @ file:///tmp/selfgz11957/pyzed-4.0-cp310-cp310-linux_x86_64.whl
pyzedhub @ file:///tmp/selfgz11957/pyzedhub-0.73-cp310-cp310-linux_x86_64.whl
pyzmq==25.1.1
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.3
radar-msgs==0.2.1
rcl-interfaces==1.2.1
rclpy==3.3.12
rcutils==5.1.5
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
resource-retriever==3.1.1
retrying==1.3.4
rich==13.6.0
rmw-dds-common==1.6.0
ros2action==0.18.9
ros2bag==0.15.9
ros2cli==0.18.9
ros2component==0.18.9
ros2doctor==0.18.9
ros2interface==0.18.9
ros2launch==0.19.7
ros2lifecycle==0.18.9
ros2multicast==0.18.9
ros2node==0.18.9
ros2param==0.18.9
ros2pkg==0.18.9
ros2run==0.18.9
ros2service==0.18.9
ros2test==0.4.0
ros2topic==0.18.9
rosapi==1.3.2
rosapi-msgs==1.3.2
rosbag2-interfaces==0.15.9
rosbag2-py==0.15.9
rosbridge-library==1.3.2
rosbridge-msgs==1.3.2
rosbridge-server==1.3.2
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpds-py==0.16.2
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.4
rqt-bag-plugins==1.1.4
rqt-console==2.0.3
rqt-graph==1.3.0
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-robot-monitor==1.0.5
rqt-runtime-monitor==1.0.0
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.12.0
sensor-msgs==4.2.3
sensor-msgs-py==4.2.3
sentry-sdk==1.32.0
setproctitle==1.3.3
shape-msgs==4.2.3
shapely==2.0.3
smclib==3.0.2
smmap==5.0.1
sniffio==1.3.0
sros2==0.10.4
stable-baselines3==2.1.0
stack-data==0.6.3
starlette==0.27.0
statistics-msgs==1.2.1
std-msgs==4.2.3
std-srvs==4.2.3
stereo-msgs==4.2.3
sympy==1.12
teleop-twist-keyboard==2.4.0
tenacity==8.2.3
tf-transformations==1.0.0
tf2-geometry-msgs==0.25.6
tf2-kdl==0.25.6
tf2-msgs==0.25.6
tf2-py==0.25.6
tf2-ros-py==0.25.6
tf2-tools==0.25.6
theora-image-transport==2.5.0
threadpoolctl==3.2.0
tifffile==2024.2.12
tokenizers==0.15.2
topic-monitor==0.20.3
topic-tools==1.1.1
topic-tools-interfaces==1.1.1
torch==2.2.2
torchvision==0.17.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
trajectory-msgs==4.2.3
transformers==4.38.1
triton==2.2.0
turtlesim==1.4.2
typing_extensions==4.8.0
tzdata==2023.3
ublox-msgs==2.3.0
udp-msgs==0.0.3
unique-identifier-msgs==2.2.1
urllib3==2.0.6
userpath==1.9.1
uvicorn==0.24.0.post1
uvloop==0.19.0
velodyne-msgs==2.4.0
virtualenv==20.25.1
visualization-msgs==4.2.3
wandb==0.15.12
watchfiles==0.21.0
wcwidth==0.2.9
Werkzeug==3.0.1
widgetsnbextension==4.0.9
xacro==2.0.8
XlsxWriter==3.2.0
zeroconf==0.32.0
